requirements.txt
django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-allauth==0.54.0
whitenoise

django-extensions
django-revproxy
gunicorn

pillow==10.0.0
python-dotenv==0.21.0
line-bot-sdk==3.2.0

# 爬蟲加載進度條
tqdm==4.64.0

matplotlib
numpy
seaborn